requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pyyaml==6.0.1
python-dateutil==2.8.2

# Let pip resolve the correct versions for these
pydantic>=2.0.0
pydantic-core>=2.0.0
typing-extensions>=4.0.0
starlette>=0.27.0
anyio>=3.7.0